+++ Makefile
@@ -1,5 +1,5 @@
 build_img:
-	docker build -t knowledge-base-generator:latest .
+	docker build -t markdown-docs-compiler:latest .
 
 lint:
 	pre-commit run --all
