+++ README.md
@@ -1,4 +1,4 @@
-knowledge-base-generator
-========================
+markdown-docs-compiler
+======================
 
 A simple utility for creating a static site from markdown files.
